requirements.txt
coloredlogs==15.0.1
flatbuffers==23.3.3
humanfriendly==10.0
mpmath==1.3.0
numpy==1.24.2
onnxruntime==1.13.1
opencv-python==4.6.0.66
packaging==23.0
pkg_resources==0.0.0
protobuf==4.22.1
PyYAML==6.0
sympy==1.11.1
